mcr-settlement-client: rework eth_client config
Rename the structs to Client and Config because this is Rust,
not some 40 year old language with modularity issues.

Make Config carry all configuration parameters for the client.
Turn build_with_urls method into build_with_config.

Make the build_with_provider method private because YAGNI.
Erase the generic transport parameter on Client because YAGNI.

+++ protocol-units/settlement/mcr/client/src/send_eth_tx.rs
@@ -65,7 +65,7 @@ pub async fn send_tx<
 >(
 	base_call_builder: CallBuilder<T, &&P, D, Ethereum>,
 	send_tx_error_rules: &[Box<dyn VerifyRule>],
-	nb_retry: usize,
+	nb_retry: u32,
 	gas_limit: u128,
 ) -> Result<(), anyhow::Error> {
 	//validate gaz price.
